backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
build==1.2.2.post1
    # via pip-tools
click==8.1.7
    # via pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
    #   tqdm
django==5.1.4
    # via
    #   django-cors-headers
    #   djangorestframework
    #   drf-spectacular
    #   drf-spectacular-sidecar
django-cors-headers==4.6.0
    # via -r requirements.in
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   drf-spectacular
drf-spectacular==0.28.0
    # via -r requirements.in
drf-spectacular-sidecar==2024.12.1
    # via -r requirements.in
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.23.0
    # via drf-spectacular
jsonschema-specifications==2024.10.1
    # via jsonschema
packaging==24.2
    # via build
pip-tools==7.4.1
    # via -r requirements.in
psycopg[binary]==3.2.3
    # via -r requirements.in
psycopg-binary==3.2.3
    # via psycopg
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pyyaml==6.0.2
    # via drf-spectacular
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
sqlparse==0.5.3
    # via django
tqdm==4.67.1
    # via -r requirements.in
typing-extensions==4.12.2
    # via psycopg
tzdata==2024.2
    # via
    #   django
    #   psycopg
uritemplate==4.1.1
    # via drf-spectacular
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
